fix(edgeless): merge view control toolbar into mouse mode toolbar
