+++ packages/blocks/src/page-block/edgeless/toolbar/edgeless-toolbar.ts
@@ -6,15 +6,23 @@ import {
   ConnectorIcon,
   HandIcon,
   ImageIcon,
+  MinusIcon,
+  PlusIcon,
   SelectIcon,
   TextIconLarge,
+  ViewBarIcon,
 } from '@blocksuite/global/config';
+import { assertExists } from '@blocksuite/global/utils';
+import { Bound, deserializeXYWH, getCommonBound } from '@blocksuite/phasor';
 import { css, html, LitElement } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 
 import type { MouseMode } from '../../../__internal__/index.js';
+import type { FrameBlockModel } from '../../../frame-block/index.js';
 import type { EdgelessPageBlockComponent } from '../edgeless-page-block.js';
 import { stopPropagation } from '../utils.js';
+
+const FIT_TO_SCREEN_PADDING = 200;
 
 @customElement('edgeless-toolbar')
 export class EdgelessToolbar extends LitElement {
@@ -48,10 +56,39 @@ export class EdgelessToolbar extends LitElement {
     .edgeless-toolbar-container[hidden] {
       display: none;
     }
+
+    .divider {
+      width: 1px;
+      height: 24px;
+      margin: 0 7px;
+      background-color: #e3e2e4;
+    }
+
+    .zoom-percent {
+      display: block;
+      box-sizing: border-box;
+      width: 48px;
+      height: 32px;
+      line-height: 22px;
+      padding: 5px;
+      border-radius: 5px;
+      font-size: 14px;
+      font-weight: 500;
+      text-align: center;
+      cursor: pointer;
+    }
+
+    .zoom-percent:hover {
+      color: var(--affine-primary-color);
+      background-color: var(--affine-hover-background);
+    }
   `;
 
   @property()
   mouseMode!: MouseMode;
+
+  @property()
+  zoom!: number;
 
   @property()
   edgeless!: EdgelessPageBlockComponent;
@@ -60,8 +97,42 @@ export class EdgelessToolbar extends LitElement {
     this.edgeless?.slots.mouseModeUpdated.emit(mouseMode);
   }
 
+  private _setZoom(zoom: number) {
+    const { viewport } = this.edgeless.surface;
+    viewport.setZoom(zoom);
+    this.edgeless.slots.viewportUpdated.emit();
+  }
+
+  private _zoomToFit() {
+    const { viewport } = this.edgeless.surface;
+    const { width, height } = viewport;
+    const frame = this.edgeless.pageModel.children[0] as FrameBlockModel;
+    const frameXYWH = deserializeXYWH(frame.xywh);
+    const frameBound = new Bound(...frameXYWH);
+
+    const surfaceElementsBound = this.edgeless.surface.getElementsBound();
+
+    const bound = surfaceElementsBound
+      ? getCommonBound([frameBound, surfaceElementsBound])
+      : frameBound;
+    assertExists(bound);
+
+    const zoom = Math.min(
+      (width - FIT_TO_SCREEN_PADDING) / bound.w,
+      (height - FIT_TO_SCREEN_PADDING) / bound.h
+    );
+
+    const cx = bound.x + bound.w / 2;
+    const cy = bound.y + bound.h / 2;
+    viewport.setZoom(zoom);
+    viewport.setCenter(cx, cy);
+    this.edgeless.slots.viewportUpdated.emit();
+  }
+
   render() {
     const type = this.mouseMode?.type;
+    const { viewport } = this.edgeless.surface;
+    const formattedZoom = `${Math.round(this.zoom * 100)}%`;
 
     return html`
       <div
@@ -114,6 +185,23 @@ export class EdgelessToolbar extends LitElement {
             this._setMouseMode({ type: 'pan', panning: false })}
         >
           ${HandIcon}
+        </edgeless-tool-icon-button>
+        <div class="divider"></div>
+        <edgeless-tool-icon-button @tool.click=${() => this._zoomToFit()}>
+          ${ViewBarIcon}
+        </edgeless-tool-icon-button>
+        <edgeless-tool-icon-button
+          @tool.click=${() => this._setZoom(viewport.zoom - 0.1)}
+        >
+          ${MinusIcon}
+        </edgeless-tool-icon-button>
+        <span class="zoom-percent" @click=${() => this._setZoom(1)}>
+          ${formattedZoom}
+        </span>
+        <edgeless-tool-icon-button
+          @tool.click=${() => this._setZoom(viewport.zoom + 0.1)}
+        >
+          ${PlusIcon}
         </edgeless-tool-icon-button>
       </div>
     `;
